[AIRFLOW-6966] Move reap_process_group to process_utils
