+++ airflow/task/task_runner/standard_task_runner.py
@@ -23,7 +23,7 @@ import psutil
 from setproctitle import setproctitle  # pylint: disable=no-name-in-module
 
 from airflow.task.task_runner.base_task_runner import BaseTaskRunner
-from airflow.utils.helpers import reap_process_group
+from airflow.utils.process_utils import reap_process_group
 
 CAN_FORK = hasattr(os, 'fork')
