+++ airflow/utils/helpers.py
@@ -16,26 +16,14 @@
 # specific language governing permissions and limitations
 # under the License.
 
-import errno
-import os
 import re
-import signal
-import subprocess
 from datetime import datetime
 from functools import reduce
 from typing import Any, Dict, Optional
 
-import psutil
 from jinja2 import Template
 
-from airflow.configuration import conf
 from airflow.exceptions import AirflowException
-
-# When killing processes, time to wait after issuing a SIGTERM before issuing a
-# SIGKILL.
-DEFAULT_TIME_TO_WAIT_AFTER_SIGTERM = conf.getint(
-    'core', 'KILLED_TASK_CLEANUP_TIME'
-)
 
 KEY_REGEX = re.compile(r'^[\w.-]+$')
 
@@ -139,84 +127,6 @@ def as_flattened_list(iterable):
     ['blue', 'red', 'green', 'yellow', 'pink']
     """
     return [e for i in iterable for e in i]
-
-
-def reap_process_group(pgid, log, sig=signal.SIGTERM,
-                       timeout=DEFAULT_TIME_TO_WAIT_AFTER_SIGTERM):
-    """
-    Tries really hard to terminate all processes in the group (including grandchildren). Will send
-    sig (SIGTERM) to the process group of pid. If any process is alive after timeout
-    a SIGKILL will be send.
-
-    :param log: log handler
-    :param pgid: process group id to kill
-    :param sig: signal type
-    :param timeout: how much time a process has to terminate
-    """
-
-    returncodes = {}
-
-    def on_terminate(p):
-        log.info("Process %s (%s) terminated with exit code %s", p, p.pid, p.returncode)
-        returncodes[p.pid] = p.returncode
-
-    def signal_procs(sig):
-        try:
-            os.killpg(pgid, sig)
-        except OSError as err:
-            # If operation not permitted error is thrown due to run_as_user,
-            # use sudo -n(--non-interactive) to kill the process
-            if err.errno == errno.EPERM:
-                subprocess.check_call(
-                    ["sudo", "-n", "kill", "-" + str(sig)] + [str(p.pid) for p in children]
-                )
-            else:
-                raise
-
-    if pgid == os.getpgid(0):
-        raise RuntimeError("I refuse to kill myself")
-
-    try:
-        parent = psutil.Process(pgid)
-
-        children = parent.children(recursive=True)
-        children.append(parent)
-    except psutil.NoSuchProcess:
-        # The process already exited, but maybe it's children haven't.
-        children = []
-        for proc in psutil.process_iter():
-            try:
-                if os.getpgid(proc.pid) == pgid and proc.pid != 0:
-                    children.append(proc)
-            except OSError:
-                pass
-
-    log.info("Sending %s to GPID %s", sig, pgid)
-    try:
-        signal_procs(sig)
-    except OSError as err:
-        # No such process, which means there is no such process group - our job
-        # is done
-        if err.errno == errno.ESRCH:
-            return returncodes
-
-    _, alive = psutil.wait_procs(children, timeout=timeout, callback=on_terminate)
-
-    if alive:
-        for proc in alive:
-            log.warning("process %s did not respond to SIGTERM. Trying SIGKILL", proc)
-
-        try:
-            signal_procs(signal.SIGKILL)
-        except OSError as err:
-            if err.errno != errno.ESRCH:
-                raise
-
-        _, alive = psutil.wait_procs(alive, timeout=timeout, callback=on_terminate)
-        if alive:
-            for proc in alive:
-                log.error("Process %s (%s) could not be killed. Giving up.", proc, proc.pid)
-    return returncodes
 
 
 def parse_template_string(template_string):
