+++ tests/utils/test_helpers.py
@@ -16,15 +16,8 @@
 # specific language governing permissions and limitations
 # under the License.
 
-import logging
-import multiprocessing
-import os
-import signal
-import time
 import unittest
 from datetime import datetime
-
-import psutil
 
 from airflow.models import TaskInstance
 from airflow.models.dag import DAG
@@ -34,33 +27,6 @@ from airflow.utils.helpers import merge_dicts
 
 
 class TestHelpers(unittest.TestCase):
-
-    @staticmethod
-    def _ignores_sigterm(child_pid, child_setup_done):
-        def signal_handler(unused_signum, unused_frame):
-            pass
-
-        signal.signal(signal.SIGTERM, signal_handler)
-        child_pid.value = os.getpid()
-        child_setup_done.release()
-        while True:
-            time.sleep(1)
-
-    @staticmethod
-    def _parent_of_ignores_sigterm(parent_pid, child_pid, setup_done):
-        def signal_handler(unused_signum, unused_frame):
-            pass
-        os.setsid()
-        signal.signal(signal.SIGTERM, signal_handler)
-        child_setup_done = multiprocessing.Semaphore(0)
-        child = multiprocessing.Process(target=TestHelpers._ignores_sigterm,
-                                        args=[child_pid, child_setup_done])
-        child.start()
-        child_setup_done.acquire(timeout=5.0)
-        parent_pid.value = os.getpid()
-        setup_done.release()
-        while True:
-            time.sleep(1)
 
     def test_render_log_filename(self):
         try_number = 1
@@ -83,35 +49,6 @@ class TestHelpers(unittest.TestCase):
         rendered_filename = helpers.render_log_filename(ti, try_number, filename_template)
 
         self.assertEqual(rendered_filename, expected_filename)
-
-    def test_reap_process_group(self):
-        """
-        Spin up a process that can't be killed by SIGTERM and make sure
-        it gets killed anyway.
-        """
-        parent_setup_done = multiprocessing.Semaphore(0)
-        parent_pid = multiprocessing.Value('i', 0)
-        child_pid = multiprocessing.Value('i', 0)
-        args = [parent_pid, child_pid, parent_setup_done]
-        parent = multiprocessing.Process(target=TestHelpers._parent_of_ignores_sigterm,
-                                         args=args)
-        try:
-            parent.start()
-            self.assertTrue(parent_setup_done.acquire(timeout=5.0))
-            self.assertTrue(psutil.pid_exists(parent_pid.value))
-            self.assertTrue(psutil.pid_exists(child_pid.value))
-
-            helpers.reap_process_group(parent_pid.value, logging.getLogger(),
-                                       timeout=1)
-
-            self.assertFalse(psutil.pid_exists(parent_pid.value))
-            self.assertFalse(psutil.pid_exists(child_pid.value))
-        finally:
-            try:
-                os.kill(parent_pid.value, signal.SIGKILL)  # terminate doesnt work here
-                os.kill(child_pid.value, signal.SIGKILL)  # terminate doesnt work here
-            except OSError:
-                pass
 
     def test_chunks(self):
         with self.assertRaises(ValueError):
